safety checkin.  creating supporting base classes to simplify code before creating more reference objects.
