+++ .gitignore
@@ -7,3 +7,4 @@ ebin
 c_src/leveldb/
 c_src/snappy*/
 c_src/system
+*~
