+++ c_src/refobjects.h
@@ -1,0 +1,231 @@
+// -------------------------------------------------------------------
+//
+// eleveldb: Erlang Wrapper for LevelDB (http://code.google.com/p/leveldb/)
+//
+// Copyright (c) 2011-2013 Basho Technologies, Inc. All Rights Reserved.
+//
+// This file is provided to you under the Apache License,
+// Version 2.0 (the "License"); you may not use this file
+// except in compliance with the License.  You may obtain
+// a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+//
+// -------------------------------------------------------------------
+
+#ifndef INCL_REFOBJECTS_H
+#define INCL_REFOBJECTS_H
+
+#include <stdint.h>
+
+#include "leveldb/db.h"
+#include "leveldb/write_batch.h"
+
+
+#ifndef INCL_THREADING_H
+    #include "threading.h"
+#endif
+
+#ifndef __WORK_RESULT_HPP
+    #include "work_result.hpp"
+#endif
+
+#ifndef ATOMS_H
+    #include "atoms.h"
+#endif
+
+
+namespace eleveldb {
+
+/**
+ * Base class for any object that offers RefInc / RefDec interface
+ */
+
+class RefObject
+{
+public:
+
+protected:
+    volatile uint32_t m_RefCount;     //!< simple count of reference, auto delete at zero
+
+public:
+    RefObject();
+
+    virtual ~RefObject();
+
+    virtual uint32_t RefInc();
+
+    virtual uint32_t RefDec();
+
+private:
+    RefObject(const RefObject&);              // nocopy
+    RefObject& operator=(const RefObject&);   // nocopyassign
+};  // class RefObject
+
+
+/**
+ * Base class for any object that is managed as an Erlang reference
+ */
+
+class ErlRefObject : public RefObject
+{
+public:
+    // these member objects are public to simplify
+    //  access by statics and external APIs
+    //  (yes, wrapper functions would be welcome)
+    volatile uint32_t m_CloseRequested;  // 1 once api close called, 2 once thread starts destructor, 3 destructor done
+
+    // DO NOT USE CONTAINER OBJECTS
+    //  ... these must be live after destructor called
+    pthread_mutex_t m_CloseMutex;        //!< for erlang forced close
+    pthread_cond_t  m_CloseCond;         //!< for erlang forced close
+
+protected:
+
+
+public:
+    ErlRefObject();
+
+    virtual ~ErlRefObject();
+
+    virtual uint32_t RefDec();
+
+private:
+    ErlRefObject(const ErlRefObject&);              // nocopy
+    ErlRefObject& operator=(const ErlRefObject&);   // nocopyassign
+};  // class RefObject
+
+
+/**
+ * Class to manage access and counting of references
+ * to a reference object.
+ */
+
+template <class TargetT>
+class ReferencePtr
+{
+    TargetT * t;
+
+public:
+    ReferencePtr()
+        : t(NULL)
+    {};
+
+    ReferencePtr(TargetT *_t)
+        : t(_t)
+    {
+        if (NULL!=t)
+            t->RefInc();
+    }
+
+    ReferencePtr(const ReferencePtr &rhs)
+    {t=rhs.t; if (NULL!=t) t->RefInc();};
+
+    ~ReferencePtr()
+    {
+        if (NULL!=t)
+            t->RefDec();
+    }
+
+    void assign(TargetT * _t) {t=_t;}; //  RefInc called by RetrieveXXObject
+
+    TargetT * get() {return(t);};
+
+    TargetT * operator->() {return(t);};
+
+private:
+ ReferencePtr & operator=(const ReferencePtr & rhs); // no assignment
+
+};  // ReferencePtr
+
+
+/**
+ * Per database object.  Created as erlang reference.
+ *
+ * Extra reference count created upon initialization, released on close.
+ */
+class DbObject : public ErlRefObject
+{
+public:
+    leveldb::DB* m_Db;                                   // NULL or leveldb database object
+
+    leveldb::Options * m_DbOptions;
+
+
+protected:
+    static ErlNifResourceType* m_Db_RESOURCE;
+
+public:
+    DbObject(leveldb::DB * DbPtr, leveldb::Options * Options);
+
+    virtual ~DbObject();  // needs to perform free_db
+
+    static void CreateDbObjectType(ErlNifEnv * Env);
+
+    static DbObject * CreateDbObject(leveldb::DB * Db, leveldb::Options * DbOptions);
+
+    static DbObject * RetrieveDbObject(ErlNifEnv * Env, const ERL_NIF_TERM & DbTerm);
+
+    static void DbObjectResourceCleanup(ErlNifEnv *Env, void * Arg);
+
+private:
+    DbObject();
+    DbObject(const DbObject&);              // nocopy
+    DbObject& operator=(const DbObject&);   // nocopyassign
+};  // class DbObject
+
+
+/**
+ * Per Iterator object.  Created as erlang reference.
+ */
+class ItrObject : public ErlRefObject
+{
+public:
+    leveldb::Iterator*   itr;
+    const leveldb::Snapshot*   snapshot;
+    bool keys_only;
+
+
+    volatile uint32_t m_handoff_atomic;  //!< matthew's atomic foreground/background prefetch flag.
+    ERL_NIF_TERM itr_ref;
+    ErlNifEnv *itr_ref_env;
+    volatile class MoveTask * reuse_move;//!< iterator work object that is reused instead of lots malloc/free
+
+    ReferencePtr<DbObject> m_DbPtr;
+
+protected:
+    static ErlNifResourceType* m_Itr_RESOURCE;
+
+public:
+    ItrObject(DbObject *, bool);
+
+    virtual ~ItrObject(); // needs to perform free_itr
+
+    static void CreateItrObjectType(ErlNifEnv * Env);
+
+    static ItrObject * CreateItrObject(DbObject * Db, bool KeysOnly);
+
+    static ItrObject * RetrieveItrObject(ErlNifEnv * Env, const ERL_NIF_TERM & DbTerm,
+                                         bool ItrClosing=false);
+
+    static void ItrObjectResourceCleanup(ErlNifEnv *Env, void * Arg);
+
+    void ReleaseReuseMove();
+
+private:
+    ItrObject();
+    ItrObject(const ItrObject &);            // no copy
+    ItrObject & operator=(const ItrObject &); // no assignment
+};  // class ItrObject
+
+} // namespace eleveldb
+
+
+#endif  // INCL_REFOBJECTS_H
